+++ src/jsMain/kotlin/dev/inmo/jsuikit/utils/Pixels.kt
@@ -1,0 +1,5 @@
+package dev.inmo.jsuikit.utils
+
+value class Pixels(val int: Int) {
+    override fun toString(): String = "${int}px"
+}