+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.0.32
 
+* Support of `UIKitHeight`
+* New builder `AttributesBuilder`
+* New `Percents`/`Pixels` abstractions
+
 ## 0.0.31
 
 * Support of `UIKitVisibility`